Rigging Lyra to receive a message from controller in editor mode.
Currently, it can ONLY receive messages in this way - the non-editor mode
doesn't work now.  The solution is to abstract the editor prep into a function
that takes arguments specifying how to proceed (e.g., whether to load up
explicit data, or the example datasets, etc.).  Then, in editor mode, the page
will wait for a message, and in non-editor mode, immediately load up the
contents.