Consider removing .jsx extensions from require() calls
By default React eslint rules flag require calls that specify `.jsx`:
the rationale from the [`eslint-plugin-react` documentation](https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/require-extension.md)
is that,

> `require()` statements should generally not include a file extension as there is a well defined mechanism for resolving a module ID to a specific file.

Our Webpack configuration does handle this resolution so it is accurate
that the extension is unneeded.

Removing the .jsx extensions from require calls suppresses 54 errors, but
I do not want to make the assumption that changing the code is the way to
go: we could also suppress this eslint rule.

There does not seem to be a universal preference about using or omitting
the extension, but the majority of coworkers I polled have adopted the
extensionless form.

Thoughts on this, @arvind @deathbearbrown?

LYRA-261