Move util.js to util/ folder
Breaking these utilities into multiple modules will enable them to be used
and tested more granularly; this will help with issues like the `require`
we use within one of the utility methods to avoid creating a circular
dependency, as each utility can be requested granularly by name once we
fragment this file. For now, the first step of that fragmentation is to
create a folder to hold these discrete modules.