Do not parse SCSS in Karma & omit js/index.js in Karma (fixes #314)
raw-loader includes the files, but includes them as raw data. The extra
file size overhead that entails isn't relevant within a testing environment,
and it gets the SCSS out of the way without impacting the JS around it.

IgnorePlugin prevents that file from being evaluated, which doesn't have any
effect on the CSS warning (that's taken care of by the raw-loader), but rather
fixes an issue where the root index.js file was being parsed and possibly causing
side-effects.

Props @mzgoddard, who helped identify that `require.context` did cause Webpack
to parse & traverse a file tree even if that file was not run through the
resulting context function. (Writing a more complex regex to omit the top-level
index file from the `require.context()` call was considered, but then abandoned
in favor of the more obvious plugin solution)