Update Issue and PR templates
The headers were registering as comments when I tried to embed these
templates directly into commit messages, and the spacing of the headers
did not accurately parse as markdown in all cases. Standardizing them
to use bold emphasis instead of headers does little to decrease the
visual priority of the header but makes it possible to use these
templates when authoring commit messages.