Add an incrementing counter utility
**Description of the problem this pull request fixes**

There are a lot of implicit incrementing counters in Lyra: this makes them
explicit and tested. As we move on to a world where we can load state from
a file or localstorage or some such, we will need to add in the ability to
seed these counters at a base starting value for each type; making a
utility sets us up to make that change easily by creating a seed or setter
method as well.

Changes proposed in this pull request:
- Implement an incrementing counter utility
- Utilize that utility within mark and primitive code

**This PR includes:**

- [x] Tests
- [x] functional comments
- [x] high level description of any new components

**Steps to functionally test this:**

- Smoke test
- Unit tests pass