Improve CancellablePromise to permit state access within executorFn
**Description of the problem this pull request fixes**

Changes proposed in this pull request:
- Modify the CancellablePromise to permit its executor function to inspect its
  cancellation state
- Build a 10ms timeout into the model parse method to permit intermediate
  cancellation without having to debounce the entire parse method
- Throttle the vega update function as a separate store listener

** If submitting code for review: **

**This PR includes:**

- [x] Tests
- [x] functional comments

**Steps to functionally test this:**

- Smoke
- Unit test suite
- Functional

A conceptual weakness of our CancellablePromise is that it did not give
itself any internal means to inspect the state of the cancellation,
precluding conditional actions inside the promise executor function.

One common scenario for which this utility was written is the Parse cycle:
imagine the following scenario,

1. User adds a mark
2. Mark registers its signals
3. User alters mark's signal values

The first two of these actions require the Vega model to be re-created;
the third requires the first two to have taken place, or the signals will
not be present when Lyra tries to set those signal values. Because mark
addition and signal registration both require a Vega re-parse, we
initialize the re-parse for each action; but we only care about the final
state, not the state after (1) alone. Exposing promise cancellation
status with a bound executor function permits the promise to be created
with a built-in debounce of 10ms, so that if the 2nd action happens
immediately after the first in execution order, we don't have to begin
to invoke the model parse at all. This is slighty more efficient than
simply throwing away the result of the parse.

The output of these debug statements:
```diff
diff --git a/src/js/model/index.js b/src/js/model/index.js
index 9764276..c4b093a 100644
--- a/src/js/model/index.js
+++ b/src/js/model/index.js
@@ -229,6 +229,7 @@ var parsePromise = null;
  * parsed and rendered.
  */
 model.parse = function(el) {
+  console.log('parsing');
   el = el || '#vis';
   if (parsePromise) {
     // A parse is already in progress; cancel that parse's callbacks
@@ -237,15 +238,19 @@ model.parse = function(el) {

   // Start the newly-requested parse within a cancellable promise
   parsePromise = new CancellablePromise(function(resolve, reject) {
+    console.log('- Starting promise');

     // Debounce parse initiation very slightly to handle re-starts on subsequent
     // store listener digest cycles: CancellablePromise exposes its state through
     // this.cancel.
     var that = this;
     setTimeout(function() {
+      console.log('  - Checking for permission to continue...');
       if (that.cancelled) {
+      console.log('    - cancelled...');
         return;
       }
+      console.log('    - ok!');
       // Recreate the vega spec
       vg.parse.spec(model.manipulators(), function(err, chart) {
         if (err) {
@@ -257,6 +262,7 @@ model.parse = function(el) {
   });

   return parsePromise.then(function(chart) {
+    console.log('      - Done!');
     model.view = chart({
       el: el
     });
@@ -373,6 +379,7 @@ function updateAllSignals(state) {

 function initiateReparse() {
   model.parse().then(function() {
+    console.log('        - done parsing');
     store.dispatch(parseInProgress(false));
   });
 }
```
is this log:
![image](https://cloud.githubusercontent.com/assets/442115/14217614/1baa5182-f81d-11e5-88d4-f3cf89b621a7.png)
You can see that of the 7 vega model re-parse requests received, this
change causes only two vega parse actions to run, and only one parse
callback to fire: the first five requests are cancelled within their
10ms debounce window, while the 6th is cancelled once the parse is in
progress but before the chart has completed and been injected into the
page.

This change allows us to remove the debounce on the parse method, cleaning
up the model code a bit. We also move to using a throttled vega update
method to avoid unnecessary updates.

Remove .only