Move all store listeners into store/listeners file and bind in js/index
Testing was becoming prohibitively difficult due to side-effects (like model
reparses) that were being generated by the listeners bound to the global
redux store. After discussion with @mzgoddard we identified two ways to improve
the situation:

1. Use the Webpack IgnorePlugin to exclude js/index.js from parsing.
2. Move the store listeners into their own file and bind them in index.js, to
   prevent listeners from being bound as a side-effect of requiring the model
   singleton (which is used widely throughout the app).

This commit accomplishes (2), above: it creates a `store/listeners.js` file,
requires that file in index.js, and uses the listeners module to create the
function which will serve as our primary store listener.