Remove a test that seems to cause a mysterious Karma error
We were getting this error in Karma:

```
02 04 2016 16:11:56.304:WARN [PhantomJS 1.9.8 (Mac OS X 0.0.0)]: Disconnected (1 times), because no message in 5000 ms.

PhantomJS 1.9.8 (Mac OS X 0.0.0): Executed 206 of 261 DISCONNECTED (5.882 secs / 0.156 secs)

02 04 2016 16:12:06.146:WARN [karma]: No captured browser, open http://localhost:9876/
```

By systematically `.skip`-ing tests in the files that have changed in
this PR, it seems that it is this test which is originating the issue.
However, I am unsatisfied that I cannot seem to deduce "why"; the test
executes successfully when run in isolation with `.only`. We are removing
it here partly because the null status of a Scale is a feature of that
constructor, not of a group per se; but if this error comes up again we
may need to reach out to the Karma team for assistance diagnosing what
it is that is making the browser disconnect and time out. (All of the
existing literature about this error originates from long build times
that cause Karma to be unable to capture the browser in the first place,
which is not the issue here.)