Fix issue where dropping a field onto a channel didn't stick
There were two major issues with dropping fields onto data: one error
around not being able to access _id of `null` was caused by the way
that the same property (fullField) was being used to hold both markup
and data, because the hideFull method was intended to null out the
markup but it was also disrupting the data used in the drop handler.

The other problem was that a method was being invoked within the rules
code without being propertly `.call`'d; the rules code is far too
dependent on context being set correctly, it could be more future-proof
to rewrite that area of the code using purer functions with less
dependence on `this`.