Update vega scene graph walker to account for Lyra structure change
When we changed the scene from type "group" to type "scene," the hierarchy
utility `.getGroupIds()` stopped returning the scene as part of its ID
array -- if the scene is no longer a group, it is only natural that it
would no logner be included. However, this broke the `.findInItemTree`
utility, which was starting its iteration at `path.length - 2` precisely
to account for the presence of the scene in the path array.

This bug manifested as the inability for selections in Lyra to propagate
correctly to Vega:

In branch lyra2:

![selection-issue-lyra2-branch](https://cloud.githubusercontent.com/assets/442115/14354226/834cab46-fcaa-11e5-917c-74582f7ebedb.gif)

In this branch:

![selection-issue-wip-branch](https://cloud.githubusercontent.com/assets/442115/14354248/90786d82-fcaa-11e5-924f-4fcdf071b987.gif)

Fixing the iteration to start at `- 1` corrects the bug. Thanks/props to
@arvind and @deathbearbrown for identifying the fix!

Issue #334 has been opened to capture unit testing this method, incidentally.