Be more efficient with function arguments within marks rules file
Anything we derive from the context does not need to be passed as an argument
as well; I could make a case that doing so muddies the fact that we are
often changing the context by reference through those arguments, rather
than explicitly assigning things in a clear way. Additionally, the way
that assignment and initialization of the new property was handled inside
bindProp obscured the workings of what is going on; doing the assignment
in one go at the end is more obvious and closer to the signal-driven form
we plan to adopt here.