Rewrite primitives reducer (and utilities and tests) to ensure Immutable objects
We have to add a lot of conditional guarding into our code if we do not
always assume that a non-primitive (boolean, number, string) value from
the immutable store will itself also be an immutable object. In order to
permit our utilities to function consistently, this PR adjusts the
primitives reducer to always set the `Immutable.fromJS` of whatever value
is provided, so that `getIn(state, 'some.path.to.arr.or.obj').toJS()` will
always be a valid call.