Do not bother setting signal value on Vega when Vega is being invalidated
The marks rule code dispatches a setProperty action, which inherently
invalidates vega because the primitives store then needs to be updated
to reflect the latest values from the redux store. If we are going to
shortly throw out the vega view anyway it is inefficient to set any
values on it.