Resolve pulse issue: Flow signal values to Lyra only through Vega
Up to now I had been pursuing a strategy by which all changes to signal
values flowed into the Lyra store and the Vega view simultaneously. Given
that the store listener called model.update but was throttled, and that
setting values on the store from within the SignalValue mixin both
dispatched singals to store AND (once model.update was fired) triggered
callbacks that themselves dispatched that same signal to the store, the
state of signals inside the store could get ambiguous and therefore the
model itself could be put into an invalid state by rapidly triggering
many signal changes one after another (e.g. scrubbing the color of a
mark in a color picker).

This adds a "dispatch" parameter to the `sg.set` method, which can be
provided to opt-out a call to `sg.set` from passing the value to the
store as well. Having this explicit opt-out claries the flow of
information within the SignalValue mixin by which we keep both Vega
_and_ the Redux store up-to-date with signal values:

**Case 1: User interacts w/ mark through handles**

1. User drags a handle
2. Vega updates the signal value internally: _Vega is now up to date_
3. Vega fires an onChange handler for that signal
4. Lyra hears that onChange event and dispatches the new signal value to
   the Redux store: _Redux is now up to date_

**Case 2: User interacts w/ mark through inspector**

1. User sets a value in the inspector
2. Lyra sets the singal value within Vega, but _not_ Redux
3. Lyra calls `model.update()`, triggering a Vega update that makes
   the new value take effect: _Vega is now up to date_
4. Vega fires an onChange handler for that signal
5. Lyra hears that onChange event and dispatches the new signal value to
   the Redux store: _Redux is now up to date_