Avoid computation other than reading from the store in mapStateToProps
This makes a small but conceptually valuable change to the work done in
`mapStateToProps`: **mapStateToProps should not be used for any purpose
beyond reading values from the store**, and **should only return values
comparable by basic equality**.

To elaborate, `mapStateToProps` fires on every digest cycle; the component's
Render will only fire when the values returned from `mapStateToProps`
change, where "change" is determined by equality. For this reason, we
should preference returning basic types from `mapStateToProps` (string,
number, boolean) and Immutable constructs, rather than objects or arrays,
because all of the former (including Immutable constructs) can be compared
by equality: any sort of `.toJS`, `.toObject`, etc, which generates a value
that cannot be compared by equality, should be handled in `.render` and not
in `mapStateToProps`, to avoid equivalent objects like `{1: true}` and
`{1: true}` being viewed as "different" values and therefore triggering
unnecessary (and therefore wasteful) re-renders.

We also only include the `ownProps` argument in `map*ToProps` when we need
it: I am not positive on this but it's possible that will reduce the
situations in which the map* method is called even further.