Extract CSS into a separate file on build
In order for Lyra to be deployed on a web server, it is most
efficient to have each of the code bundles separated out into
scripts or styles as appropriate.

This PR alters our `npm run build` configuration to render all
SASS content into a single style.css file at the end of the build
process.

To test this, run `npm run build`, then use a static file server
to serve the application from the repo root directory (do not use
`npm start`); everything should work as it does when using the
dev server.

Note that this commit removes the plugin that was being used to
copy the index HTML into the build directory; the index.html from
the repository root is the one with the correct paths for use with
built assets.