Add a production-mode build to Lyra
This adds a `npm run build:prod` script, which uses a new production-mode
wrapper for the webpack config to set certain optimization options useful
when deploying Lyra as a standalone app on a static file server.

The key changes in a production-mode build:

1. Use of the [`-p` Webpack config option](https://webpack.github.io/docs/cli.html#production-shortcut-p) to enable Webpack's build optimizations
2. Use of a different sourcemap option
3. Use of a DefinePlugin to globally hard-set a "production mode" constant to permit Redux and other code to opt-in to performance-tuned builds instead of development-tool builds

With this change, a complete "deployment" workflow will look like this:

1. `npm run build:prod` to trigger a build in production mode
2. Copy `index.html`, `build/`, and `data/` to a server
3. Lyra should now be usable on that server

Partially addresses #362 by providing tools for generating a truly standalone
build: webpack-dev-server was never intended for production use.