Refactor dataset-utils as PipelineModal now loads+parses values.
TODO: In this commit, we still store both the raw and parsed values.
This unnecessarily increases Lyra's footprint. Raw values, however,
are useful when exporting the Vega JSON spec as, in the case of CSV
and TSV, can be more concise than the parsed JSON. Consider
converting parsed JSON back to CSV/TSV on export.