Facet line/area marks when binding to color.
Unlike Lyra 1, we hide the fact that line/area marks need to be nested
within a faceted group as this feels like a Vega-level issue that Lyra
users should not need to worry about. Instead, we store facets as a
hidden property of Mark records, and deal with them during export.